requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.7.4
openai==1.47.0
aiogram==3.13.0
pdfplumber==0.11.4
PyPDF2==3.0.1
python-multipart==0.0.12
pytz==2024.2
pytest==8.3.3
pytest-asyncio==0.24.0
